Parse catalog quantities from names
